requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2
alembic
python-dotenv
pydantic
redis
passlib[bcrypt]
